requirements.txt
# Requirements for PetLoveHealth Edge App
paho-mqtt==1.6.1